chore: new mapper for generator set process unit and move resource to domain
